+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 * Fix problem with [HexaPDF::Filter::Predictor] due to behaviour change of Ruby
   2.6.0 in `String#setbyte`
 * Fitting of [HexaPDF::Layout::TextBox] when the box has padding and/or borders
+* Fitting of absolutely positioned boxes in [HexaPDF::Layout::Frame]
 
 ## 0.8.0 - 2018-10-26